[args] Add a new JSON type for arguments
This new type can be used in the arguments struct to be able to pass
more complex structures from the command line or the environment.

When parsing, this type expects a serialized JSON string which will be
converted to a generic map using the JSON type in the args package.