Downcase import line for logrus
Here we downcase import line for logrus to avoid an issue where go get tries to
include a package by the same name, but lowercased and fails.  This comes up as
part of including logrus in another project, where we want to use this SDK and
can't duplicate the logrus package with capitalizing our import.

I think the general case here is to have the package name downcased, even though
github works with either.