Update integration.Publish method to print breaklines
The contract between the agent and integrations is that integrations
would print a single line output. If the execution of the integration
binary returns multiple integration outputs, they have to be separated
by a breakline.

You can compare the current behavior with v2.2.0:

https://github.com/newrelic/infra-integrations-sdk/blob/v2.2.0/sdk/data.go#L95

As you can see, in that version we were using `Println()`