Stop reading from stdOut after receiving message
Previously, `doQuery` was only returning when an EOF was encountered.
However, this causes issues when multiple queries are to be sent to the
`nrjmx` process. Multiple `doQuery` goroutines were running at the
same time, and it appears the first one to run was always consuming
stdOut, which caused all remaining queries to fail with a timeout.